Update llmcord.py
imported commands from discord.py to allow status changes depending on endpoint API connection failure -- if the api endpoint fails, the discord bot will 'go offline'. 

additionally, includes a small adjustment to the prompt related to the injection of the date, to attempt to dissuade the llm from referring to it when it is unsure of what to say, as this comes off strange.

finally, adds gemma3 as a list of vision-enabled models.